app/auth/requirements.txt
sqlmodel
psycopg2-binary
argon2-cffi
pyjwt
fastapi
uvicorn[standard]
